remove log config args and suggest setting ENV

+++ ngtypes/difficulty.go
@@ -12,7 +12,7 @@ func GetNextDiff(tailBlock *Block) *big.Int {
 		return diff
 	}
 
-	elapsed := int64(tailBlock.Timestamp - GenesisTimestamp)
+	elapsed := tailBlock.Timestamp - GenesisTimestamp
 	elapsed = elapsed - int64(tailBlock.GetHeight())*int64(TargetTime/time.Second)
 	delta := new(big.Int)
 	if elapsed < int64(TargetTime/time.Second)*(-2) {
